Rename operator package to traversal
Resolves #41
