+++ src/main/kotlin/org/treeWare/model/cursor/LeaderModelCursor.kt
@@ -1,9 +1,9 @@
 package org.treeWare.model.cursor
 
 import org.treeWare.model.core.*
-import org.treeWare.model.operator.AbstractLeader1Follower0ModelVisitor
-import org.treeWare.model.operator.TraversalAction
-import org.treeWare.model.operator.dispatchVisit
+import org.treeWare.model.traversal.AbstractLeader1Follower0ModelVisitor
+import org.treeWare.model.traversal.TraversalAction
+import org.treeWare.model.traversal.dispatchVisit
 import java.util.*
 
 class LeaderModelCursor<Aux>(private val initial: ElementModel<Aux>) {
